Add additional expiring test case and handle unknown template type
test expiration for the following
1. assume role statement
2. inline statement
3. managed policy

I change the behavior to continue if we re-counter unknown template
type. Previously, we raise immediately after logging the error.
However, we really should not stop the whole process if we encounter
an unknown template type. Templates can be ahead or behind what
iambic binary version we are using.